+++ server/component/moduleQualtricsProjectAction/ModuleQualtricsProjectActionView.php
@@ -316,7 +316,7 @@ class ModuleQualtricsProjectActionView extends ModuleQualtricsProjectView
                     "id" => "config",
                     "name" => "schedule_info[config]",
                     "value" => isset($this->action["schedule_info"]['config']) ? $this->action["schedule_info"]['config'] : '',
-                    "css" => "mb-3",
+                    "css" => "mb-3 actionConfig",
                     "placeholder" => "",
                 ))
             )