+++ server/component/style/textarea/TextareaView.php
@@ -143,14 +143,58 @@ class TextareaView extends FormFieldView
                                     "label" => "Save",
                                     "url" => "#",
                                     "type" => "primary",
-                                    "css" => "savequaltricsSurveyConfigBuilder"
+                                "css" => "savequaltricsSurveyConfigBuilder"
+                            )),
+                        )
+                    ))
+                ),
+            ));
+            $modal->output_content();
+            require __DIR__ . "/tpl_qualtricsSurveyConfig_builder.php";
+        } else if (strpos($this->css, 'actionConfig') !== false) {
+            $modal = new BaseStyleComponent('modal', array(
+                'title' => "Action Config Builder",
+                "css" => "actionConfig_builder_modal_holder",
+                'children' => array(
+                    new BaseStyleComponent("div", array(
+                        "css" => "actionConfig_builder"
+                    )),
+                    new BaseStyleComponent("div", array(
+                        "css" => "modal-footer",
+                        "children" => array(
+                            new BaseStyleComponent("button", array(
+                                "label" => "Save",
+                                "url" => "#",
+                                "type" => "primary",
+                                "css" => "saveActionConfigBuilder"
+                            )),
+                        )
+                    ))
+                ),
+            ));
+            $modal->output_content();
+            $modalCondition = new BaseStyleComponent('modal', array(
+                    'title' => "Action Condition Builder",
+                    "css" => "action_condition_builder_modal_holder",
+                    'children' => array(
+                        new BaseStyleComponent("div", array(
+                            "css" => "action_condition_builder"
+                        )),
+                        new BaseStyleComponent("div", array(
+                            "css" => "modal-footer",
+                            "children" => array(
+                                new BaseStyleComponent("button", array(
+                                    "label" => "Save",
+                                    "url" => "#",
+                                    "type" => "primary",
+                                    "css" => "saveActionConditionBuilder"
                                 )),
                             )
                         ))
                     ),
                 ));
-            $modal->output_content();
-            require __DIR__ . "/tpl_qualtricsSurveyConfig_builder.php";
+            $modalCondition->output_content();
+            require __DIR__ . "/tpl_actionConfig_builder.php";
         }
     }
 }