+++ server/service/conditions/Condition.php
@@ -1,0 +1,126 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<?php
+require_once __DIR__ . "/JsonLogic.php";
+/**
+ * This class allows to check different condtions based on JSON logic.
+ */
+class Condition
+{
+    /**
+     * The db instance which grants access to the DB.
+     */
+    private $db;
+
+    /**
+     * The db instance which grants access to the user input.
+     */
+    private $user_input;
+
+    /**
+     * Start the session.
+     *
+     * @param object $db
+     *  The db instance which grants access to the DB.
+     * @param object $user_input
+     *  The user_input instance which grants access to the user_input.
+     */
+    public function __construct($db, $user_input)
+    {
+        $this->db = $db;
+        $this->user_input = $user_input;
+    }
+
+    /**
+     * Check if the logged user is in this group.
+     *
+     * @param string $groupName
+     *  group name that we want to check
+     * @param int $id_users
+     * the user who we will check if is in the group
+     * @retval boolean
+     *  returns true if the user belnogs to the group and false if not
+     */
+    private function get_user_group($groupName, $id_users)
+    {
+        $sql = "select g.name as group_name
+                from users u
+                inner join users_groups ug on (u.id = ug.id_users)
+                inner join groups g on (ug.id_groups = g.id)
+                where g.name = :group and u.id = :uid";
+        $res = $this->db->query_db_first($sql, array(
+            ':group' => $groupName,
+            ':uid' => $id_users
+        ));
+        return  isset($res['group_name']);
+    }
+
+    /**
+     * Use the JsonLogic libarary to compute whether the json condition is true
+     * or false.
+     *
+     * @param array $condition
+     *  An array representing the json condition string.
+     * @param int $id_users
+     * the user who we will check if is in the group. If not set then the session logged user
+     * @param string $section
+     * the name of the section that we checked, if it is not form section then we specify
+     * @retval mixed
+     *  The evaluated condition.
+     */
+    public function compute_condition($condition, $id_users = null, $section = 'system')
+    {
+        $id_users = $id_users ? $id_users : $_SESSION['id_user']; // set default value to $_SESSION['id_user'] if not set
+        $res = array("result" => false, "fields" => array());
+        if ($condition === null || $condition === "")
+            return true;
+        $j_condition = json_encode($condition);
+        $j_condition = str_replace('__current_date__', date('Y-m-d'), $j_condition); // replace __current_date__
+        $j_condition = str_replace('__current_date_time__', date('Y-m-d H:i'), $j_condition); // replace __current_date_time__
+        $j_condition = str_replace('__current_time__', date('H:i'), $j_condition); // replace __current_time__
+        // replace form field keywords with the actual values.
+        $pattern = '~"' . $this->user_input->get_input_value_pattern() . '"~';
+        preg_match_all($pattern, $j_condition, $matches, PREG_PATTERN_ORDER);
+        foreach ($matches[0] as $match) {
+            $val = $this->user_input->get_input_value_by_pattern(trim($match, '"'), $id_users);
+            if ($val === null) {
+                $res['fields'][$match] = "bad field syntax";
+            }
+            else if ($val === "") {
+                $res['fields'][$match] = "no value stored for this field";
+            }
+            else {
+                $res['fields'][$match] = $val;
+            }
+            $j_condition = str_replace($match, '"' . $val . '"', $j_condition);
+        }
+
+        preg_match_all('~"\$[^"@#]+"~', $j_condition, $matches, PREG_PATTERN_ORDER); // group pattern
+        foreach ($matches[0] as $match) {
+            $groupName = trim($match, '"');
+            $groupName = str_replace("$", "", $groupName);
+            $val = $this->get_user_group($groupName, $id_users);
+            $res['group_name'] = $val;
+            $j_condition = str_replace($match, '"' . $val . '"', $j_condition);
+        }
+        // compute the condition
+        try
+        {
+            $j_condition = str_replace("\\", '\\\\', $j_condition); //get rid of new lines, otherwise it fails
+            $j_condition = str_replace("\r\n", '\n', $j_condition); //get rid of new lines, otherwise it fails
+            $j_condition = json_decode($j_condition, true);
+            $res['result'] = JsonLogic::apply($j_condition);
+        }
+        catch(\Exception | \ArgumentCountError $e)
+        {
+            $res['fields'] = "JsonLogic::apply() failed in section '"
+                . $section . "': " . $e->getMessage();
+        }
+
+        return $res;
+    }
+}
+?>